Remove Advent/Weekly/Monthly calendars
